moved api/business logic into seperate files
