+++ src/views/pages/home.ejs
@@ -6,6 +6,7 @@
   <div class="container">
     <!-- THE REST OF YOUR HTML CODE GOES HERE -->
     <p>Home page</p>
+    <p>Welcome <%- user.username %></p> 
   </div>
 </main>
