+++ src/views/partials/menu.ejs
@@ -15,7 +15,7 @@
           <ul class="navbar-nav me-auto">
             <li class="nav-item">
               <!-- TO-DO: For Discover, add a <a> tag with an attribute href to call the '/discover' API -->
-                <a class="nav-link" href="/home">Home</a>
+                <a class="nav-link" href="/user/home">Home</a>
             </li>
           </ul>
           <div class="nav-item me-1">
